+++ examples/clash_handling.rs
@@ -44,7 +44,7 @@ fn spawn_input_map(mut commands: Commands) {
         ButtonlikeChord::new([Digit1, Digit2, Digit3]),
     );
 
-    commands.spawn(InputManagerBundle::with_map(input_map));
+    commands.spawn(input_map);
 }
 
 fn report_pressed_actions(