+++ examples/input_processing.rs
@@ -47,9 +47,7 @@ fn spawn_player(mut commands: Commands) {
                 DualAxisSensitivity::all(2.0).into(),
             ]),
         );
-    commands
-        .spawn(InputManagerBundle::with_map(input_map))
-        .insert(Player);
+    commands.spawn(input_map).insert(Player);
 }
 
 fn check_data(query: Query<&ActionState<Action>, With<Player>>) {