+++ examples/multiplayer.rs
@@ -23,13 +23,7 @@ enum Player {
     Two,
 }
 
-#[derive(Bundle)]
-struct PlayerBundle {
-    player: Player,
-    input_manager: InputManagerBundle<Action>,
-}
-
-impl PlayerBundle {
+impl Player {
     fn input_map(player: Player, gamepad_0: Entity, gamepad_1: Entity) -> InputMap<Action> {
         let mut input_map = match player {
             Player::One => InputMap::new([
@@ -67,23 +61,15 @@ fn spawn_players(mut commands: Commands) {
     let gamepad_0 = commands.spawn(()).id();
     let gamepad_1 = commands.spawn(()).id();
 
-    commands.spawn(PlayerBundle {
-        player: Player::One,
-        input_manager: InputManagerBundle::with_map(PlayerBundle::input_map(
-            Player::One,
-            gamepad_0,
-            gamepad_1,
-        )),
-    });
+    commands.spawn((
+        Player::One,
+        Player::input_map(Player::One, gamepad_0, gamepad_1),
+    ));
 
-    commands.spawn(PlayerBundle {
-        player: Player::Two,
-        input_manager: InputManagerBundle::with_map(PlayerBundle::input_map(
-            Player::Two,
-            gamepad_0,
-            gamepad_1,
-        )),
-    });
+    commands.spawn((
+        Player::Two,
+        Player::input_map(Player::Two, gamepad_0, gamepad_1),
+    ));
 }
 
 fn move_players(player_query: Query<(&Player, &ActionState<Action>)>) {