+++ examples/single_player.rs
@@ -57,15 +57,7 @@ impl ArpgAction {
 #[derive(Component)]
 pub struct Player;
 
-#[derive(Bundle)]
-struct PlayerBundle {
-    player: Player,
-    // This bundle must be added to your player entity
-    // (or whatever else you wish to control)
-    input_manager: InputManagerBundle<ArpgAction>,
-}
-
-impl PlayerBundle {
+impl Player {
     fn default_input_map() -> InputMap<ArpgAction> {
         // This allows us to replace `ArpgAction::Up` with `Up`,
         // significantly reducing boilerplate
@@ -108,10 +100,7 @@ impl PlayerBundle {
 }
 
 fn spawn_player(mut commands: Commands) {
-    commands.spawn(PlayerBundle {
-        player: Player,
-        input_manager: InputManagerBundle::with_map(PlayerBundle::default_input_map()),
-    });
+    commands.spawn((Player, Player::default_input_map()));
 }
 
 fn cast_fireball(query: Query<&ActionState<ArpgAction>, With<Player>>) {